refactored to ms with rest api -- still some places will require cleanup in the future
