+++ package.json
@@ -17,17 +17,24 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "-": "0.0.1",
     "@doc-gen/dg-data-provider-azuredevops": "0.0.5",
-    "@doc-gen/dg-skins": "0.0.1",
+    "@doc-gen/dg-skins": "0.0.3",
     "axios": "^0.21.0",
     "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-prom-bundle": "^6.4.1",
     "html-table-to-json": "^1.0.0",
+    "http-status-codes": "^2.1.4",
     "node-fetch": "^2.6.1",
+    "prom-client": "^14.0.1",
     "request": "^2.88.2",
+    "save": "^2.4.0",
     "string-strip-html": "^6.1.1",
     "winston": "^3.2.1"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
     "@types/jest": "^25.2.3",
     "@types/node": "^13.13.5",
     "jest": "^26.6.3",
